+++ src/main/kotlin/com/assari/voicebooklm/config/ExternalClientConfig.kt
@@ -2,8 +2,8 @@ package com.assari.voicebooklm.config
 
 import com.assari.voicebooklm.infrastructure.api.ai.GeminiAiMemoFormatter
 import com.assari.voicebooklm.infrastructure.api.speech.GoogleSpeechTranscriber
-import com.assari.voicebooklm.usecase.memo.port.AiMemoFormatter
-import com.assari.voicebooklm.usecase.memo.port.SpeechTranscriber
+import com.assari.voicebooklm.usecase.memo.client.AiMemoFormatter
+import com.assari.voicebooklm.usecase.memo.client.SpeechTranscriber
 import com.google.api.gax.core.FixedCredentialsProvider
 import com.google.auth.oauth2.GoogleCredentials
 import com.google.cloud.speech.v1.SpeechClient
