+++ src/main/kotlin/com/assari/voicebooklm/infrastructure/api/speech/GoogleSpeechTranscriber.kt
@@ -1,8 +1,8 @@
 package com.assari.voicebooklm.infrastructure.api.speech
 
-import com.assari.voicebooklm.usecase.memo.port.SpeechTranscriber
-import com.assari.voicebooklm.usecase.memo.port.SpeechTranscription
-import com.assari.voicebooklm.usecase.memo.port.SpeechTranscriptionCommand
+import com.assari.voicebooklm.usecase.memo.client.SpeechTranscriber
+import com.assari.voicebooklm.usecase.memo.client.SpeechTranscription
+import com.assari.voicebooklm.usecase.memo.client.SpeechTranscriptionCommand
 import com.google.cloud.speech.v1.RecognitionAudio
 import com.google.cloud.speech.v1.RecognitionConfig
 import com.google.cloud.speech.v1.SpeechClient
