+++ src/main/kotlin/com/assari/voicebooklm/infrastructure/postgres_jpa/memo/MemoJpaDataRepository.kt
@@ -2,6 +2,7 @@ package com.assari.voicebooklm.infrastructure.postgres_jpa.memo
 
 import java.util.UUID
 import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.data.jpa.repository.Modifying
 import org.springframework.stereotype.Repository
 
 /** MemoJpaEntity 用 Spring Data JPA リポジトリ。 */
@@ -10,4 +11,7 @@ interface MemoJpaDataRepository : JpaRepository<MemoJpaEntity, UUID> {
     fun findByIdAndDeletedFalse(id: UUID): MemoJpaEntity?
 
     fun findByUserIdAndDeletedFalseOrderByCreatedAtDesc(userId: UUID): List<MemoJpaEntity>
+
+    @Modifying
+    fun deleteByUserId(userId: UUID)
 }
