+++ src/main/kotlin/com/assari/voicebooklm/infrastructure/postgres_jpa/memo/MemoRepositoryImpl.kt
@@ -1,13 +1,32 @@
 package com.assari.voicebooklm.infrastructure.postgres_jpa.memo
 
+import com.assari.voicebooklm.domain.model.Memo
+import com.assari.voicebooklm.domain.repository.MemoRepository
 import java.util.UUID
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
-/** MemoJpaEntity 用 Spring Data JPA リポジトリ。 */
+/**
+ * MemoRepository の JPA 実装。
+ */
 @Repository
-interface MemoJpaDataRepository : JpaRepository<MemoJpaEntity, UUID> {
-    fun findByIdAndDeletedFalse(id: UUID): MemoJpaEntity?
+class MemoRepositoryImpl(
+    private val memoJpaRepository: MemoJpaDataRepository,
+) : MemoRepository {
 
-    fun findByUserIdAndDeletedFalseOrderByCreatedAtDesc(userId: UUID): List<MemoJpaEntity>
+    override suspend fun save(memo: Memo): Memo {
+        val entity = MemoJpaEntity.fromDomain(memo)
+        val saved = memoJpaRepository.save(entity)
+        return saved.toDomain()
+    }
+
+    override suspend fun findById(id: UUID): Memo? =
+        memoJpaRepository.findByIdAndDeletedFalse(id)?.toDomain()
+
+    override suspend fun findByUserId(userId: UUID): List<Memo> =
+        memoJpaRepository.findByUserIdAndDeletedFalseOrderByCreatedAtDesc(userId).map { it.toDomain() }
+
+    override fun deleteByUserId(userId: UUID) {
+        memoJpaRepository.deleteByUserId(userId)
+    }
 }
