+++ src/test/kotlin/com/assari/voicebooklm/presentation/controller/voice/VoiceControllerTest.kt
@@ -1,13 +1,13 @@
 package com.assari.voicebooklm.presentation.controller.voice
 
-import com.assari.voicebooklm.domain.model.memo.Memo
+import com.assari.voicebooklm.domain.model.Memo
 import com.assari.voicebooklm.presentation.exception.ErrorResponse
 import com.assari.voicebooklm.usecase.memo.CreateMemoCommand
 import com.assari.voicebooklm.usecase.memo.CreateMemoResult
 import com.assari.voicebooklm.usecase.memo.CreateMemoUseCase
 import com.assari.voicebooklm.usecase.memo.FallbackUsage
 import com.assari.voicebooklm.usecase.memo.ProcessingTime
-import com.assari.voicebooklm.usecase.memo.port.SpeechTranscription
+import com.assari.voicebooklm.usecase.memo.client.SpeechTranscription
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.mockk
